requirements.txt
customtkinter
pyautogui
pynput
keyboard
pillow
